# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/models": {
      "get": {
        "summary": "获取模型列表",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "系统"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/mj/": {
      "get": {
        "summary": "获取所有Midjourney任务",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "任务"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/models/": {
      "get": {
        "summary": "获取所有模型元数据",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "模型管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "post": {
        "summary": "创建模型元数据",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "模型管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "put": {
        "summary": "更新模型元数据",
        "deprecated": false,
        "description": "👨‍💼 需要管理员权限（Admin）",
        "tags": [
          "模型管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken1": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "new_api_refresh",
        "description": "仅 refresh/logout 使用的 HttpOnly、SameSite=Strict Cookie"
      },
      "SecurityProof": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Security-Proof",
        "description": "绑定用户、登录会话和 scope 的短期二次验证 Proof"
      },
      "Combination222": {
        "group": [
          {
            "id": 573667
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}